services/python-backend/requirements.txt
# Базовые зависимости
fastapi[all]
numpy>=1.21.0,<2.0.0
pillow>=10.0.0
tqdm>=4.66.0
pandas>=2.0.0
tiktoken>=0.5.0
matplotlib>=3.7.0
einops>=0.7.0
albumentations
timm
scikit-learn


# Специфичные для Qwen2.5-VL
qwen-vl-utils[decord]==0.0.8  # Для обработки мультимодальных данных
transformers @ git+https://github.com/huggingface/transformers  # Обязательно из исходников!
accelerate>=0.27.0  # Свежая версия

# Torch с поддержкой CUDA 12.1
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.2.0+cu121
torchvision==0.17.0+cu121
torchaudio==2.2.0+cu121

# Дополнительно (если нужно квантование)
bitsandbytes>=0.43.0  # Только если планируете 4-битную загрузку
